Rework SpeziSecureStorage API (#32)
# Rework SpeziSecureStorage API

## :recycle: Current situation & Problem
- SpeziStorage currently consists of two targets: SpeziLocalStorage and
SpeziSecureStorage
- The names are misleading; the fact that one of them is called the
"secure" storage might lead people to assume that the other one is not
secure. This is false. They both are.
- → we rename `SecureStorage` to `CredentialsStorage`, to make its name
match more closely its intended purpose, and also get rid of any
potential for misunderstanding the individual modules' intended purposes
- The fact that the `CredentialsStorage` API is using raw `String`
values to identify entries in the keychain, and is making heavy use of
optional parameters with default `nil` values, makes it very easy to
mess up operations that are intended to operate on the same entry in the
keychain (since you need to remember to always pass in the exact same
set of parameters in all your callsites, which is very easy to get
wrong)
- in fact, there are some bugs in the current (ie, pre-PR)
implementation which are a result of nil-defaulted function parameters
not being specified when they should be given an explicit value.
- → we rework the API a little bit, to introduce dedicated `KeyTag` (for
working with keys) and `CredentialsStorageKey` (for working with
credentials) types, which are used to identify, in a more safe manner,
the entries an operation should work with.
- The `CredentialsStorageKey`, for example, combines all information
which is required to correctly define how a credentials entry is /
should be stored in the keychain, instead of requiring the user to
remember to always pass in the correct same values whenever they want to
store/retrieve/delete a credentials entry.


## :gear: Release Notes 
- Renamed and reworked `SecureStorage` into `CredentialsStorage`
- Introduced `KeyTag` type to identify keys
- Introduced `CredentialsStorageKey` type to define credentials entries
in the `CredentialsStorage`.


## :books: Documentation
The documentation has been updated to reflect the new structure of the
API.


## :white_check_mark: Testing
The tests have been updated to work with the new API.
Since the functionality of the API is unchanged, there are no new tests.


## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md).

---------

Co-authored-by: Paul Schmiedmayer <PSchmiedmayer@users.noreply.github.com>